API/requirements.txt
fastapi
uvicorn[standard]
keras
joblib
pandas
numpy
pyserial
shutil
